backend/requirements.txt
aiohttp
trafilatura
beautifulsoup4
python-dateutil
langdetect
selenium
playwright
tqdm
nltk
tiktoken
python-dotenv
openai
fastapi
uvicorn[standard]
python-multipart
sqlalchemy[asyncio]
asyncpg
pgvector
pydantic
